# pozoristeback/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Teatar Maska API",
        "description": "API dokumentacija za seminarski rad",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Lokalni server"
        }
    ],
    "paths": {
...
        "/api/sale": {
            "get": {
                "tags": [
                    "Sale"
                ],
                "summary": "Prikaz svih sala",
                "description": "Display a listing of the resource.",
                "operationId": "6d4b645af564be050ee632110762ec82",
                "responses": {
                    "200": {
                        "description": "Uspešno učitane sale"
                    }
                }
            },
            "post": {
                "tags": [
                    "Sale"
                ],
                "summary": "Dodavanje nove sale",
                "description": "Store a newly created resource in storage.",
                "operationId": "7bfbfd762f0b9cc4165ecd123857e0ca",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "naziv",
                                    "kapacitet"
                                ],
                                "properties": {
                                    "naziv": {
                                        "type": "string",
                                        "example": "Velika scena"
                                    },
                                    "kapacitet": {
                                        "type": "integer",
                                        "example": 350
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Uspešno kreirano"
                    },
                    "422": {
                        "description": "Greška u validaciji"
                    }
                }
            }
        },
        "/api/sale/{id}": {
            "get": {
                "tags": [
                    "Sale"
                ],
                "summary": "Prikaz jedne sale",
                "description": "Display the specified resource.",
                "operationId": "35fe23d1ad40587f8fc1d3450da074f2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Uspešno"
                    },
                    "404": {
                        "description": "Sala nije pronađena"
                    }
                }
            },
            "put": {
                "tags": [
                    "Sale"
                ],
                "summary": "Ažuriranje sale",
                "description": "Update the specified resource in storage.",
                "operationId": "931d0232052188c0491eca29e6c765d1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "naziv": {
                                        "type": "string",
                                        "example": "Mala scena - renovirana"
                                    },
                                    "kapacitet": {
                                        "type": "integer",
                                        "example": 120
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Uspešno ažurirano"
                    },
                    "404": {
                        "description": "Sala nije pronađena"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Sale"
                ],
                "summary": "Brisanje sale",
                "description": "Remove the specified resource from storage.",
                "operationId": "567c3e7bed0d1d0f118d231a4bb677d7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Obrisano"
                    },
                    "404": {
                        "description": "Sala nije pronađena"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Unesite vaš API token koji ste dobili pri prijavi (login) ili registraciji.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}